requirements.txt
pdfplumber>=0.11.3
pandas
google-genai
python-dotenv
pdf2image
